parsl-containers/ward_bmg/requirements.txt
gzip
pickle
matplotlib
python-ternary
pymatgen
git://github.com/WardLT/matminer.git@sciadv_features#egg=matminer